+++ src/main/java/kr/santanrudolph/everyvent/domain/calendar/repository/ScrapCountProjection.java
@@ -1,4 +1,4 @@
-package kr.santanrudolph.everyvent.domain.calendar.scrap;
+package kr.santanrudolph.everyvent.domain.calendar.repository;
 
 public interface ScrapCountProjection {
   Long getCalendarId();
